+++ checkpoint/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - #v1 Add Save option `enable_replica_parallel_separate_folder`
+- Add support for dispatchers in Jax Array Handlers.
 
 ### Changed